Can now generate very large images, with good smooth colors.

+++ mandelbrot.h
@@ -2,6 +2,6 @@
 #define _MANDELBROT_H_
 #include <stdio.h>
 
-int *create_mandelbrot(int res_x, int res_y, float min_x, float min_y,
+float *create_mandelbrot(int res_x, int res_y, float min_x, float min_y,
         float max_x, float max_y, int iter);
 #endif
